+++ pages/site_sidebar.php
@@ -81,6 +81,7 @@
         <li><a href="cdc_dashboard.php"><i class="fa fa-pie-chart"></i> Dashboard</a></li>
         <li><a href="cdc_organismo.php"><i class="fa fa-building"></i> Organismos</a></li>
         <li><a href="cdc_cliente.php"><i class="fa fa-user"></i> Clientes</a></li>
+        <li><a href="cdc_servidores.php"><i class="fa fa-server"></i> Servidores</a></li>
         <li class="treeview">
           <a href="#"><i class="fa fa-gears"></i> Servicios
             <span class="pull-right-container">